Normalize Python exception values before processing them
An exception value returned by PyErr_Fetch() can be in certain
situations "unnormalized", which means the value does not have to be an
instance of the returned exception type. This typically happens when an
exception is set from a native library. For instance:
> PyObject *w = Py_BuildValue("(i,i)", 123, 456);
> PyErr_SetObject(NativeError, w);

The exception value is in this case a tuple object and not an instance
of NativeError, until the exception gets normalized.

Code in YPython::PyErrorHandler() passes an obtained exception value
directly to traceback.format_exception(). This method however expects
that the value is already a valid exception object and fails quickly
with an AttributeError when that is not the case. This subsequently
results in an unhandled error in YPython::PyErrorHandler() and in a
SystemError when other Python/C API functions are invoked.

The patch fixes the problem by calling PyErr_NormalizeException() in
YPython::PyErrorHandler() to normalize the value to be an exception
object.